--- apps/dev/sveltekit/src/routes/protected-ssr-login/+page.server.ts
@@ -1,13 +1,0 @@
-import type { LayoutServerLoad } from "../../$types"
-
-export const load: LayoutServerLoad = async (event) => {
-  const session = await event.locals.auth()
-
-  if (!session) {
-    await event.locals.signIn()
-  }
-
-  return {
-    session,
-  }
-}
