+++ apps/dev/sveltekit/src/routes/sign-in/+page.server.ts
@@ -1,0 +1,3 @@
+import { signIn } from "../../auth"
+import type { Actions } from "./$types"
+export const actions = { default: signIn } satisfies Actions
