+++ packages/frameworks-sveltekit/src/lib/components/SignIn.svelte
@@ -1,0 +1,29 @@
+<script lang="ts">
+  import type { signIn } from "$lib/actions"
+
+  export let className = ""
+  export let provider: Parameters<typeof signIn>[0]
+  export let callbackUrl: string = '/'
+  export let signInPage = "sign-in"
+  export let options: Parameters<typeof signIn>[1] | undefined = undefined
+  export let authorizationParams: Parameters<typeof signIn>[2] | undefined = undefined
+</script>
+
+<form
+  method="POST"
+  action={`/${signInPage}${
+    authorizationParams && Object.keys(authorizationParams).length
+      ? `?${new URLSearchParams(authorizationParams).toString()}`
+      : ""
+  }`}
+  class={`signInButton ${className}`}
+>
+  <input type="hidden" name="id" value={provider} />
+  <input type="hidden" name="callbackUrl" value={callbackUrl} />
+  {#if options}
+    {#each Object.entries(options) as [key, value]}
+      <input type="hidden" name={key} value={value} />
+    {/each}
+  {/if}
+  <button type="submit"><slot /></button>
+</form>
