+++ packages/frameworks-sveltekit/src/lib/components/index.ts
@@ -1,0 +1,3 @@
+import SignIn from "./SignIn.svelte"
+
+export { SignIn }
